Inclusão de exceptions e arquivo init.js para popular tabelas com doc…

+++ src/main/java/com/tech_challenge_fiap/adapter/service/outbound/repository/ProductRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.tech_challenge_fiap.adapter.service.outbound.entity.ProductEntity;
 
 import com.tech_challenge_fiap.core.domain.product.ProductRepository;
 import com.tech_challenge_fiap.util.Enum.CategoryEnum;
+import com.tech_challenge_fiap.util.exception.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
@@ -18,26 +19,49 @@ public class ProductRepositoryImpl implements ProductRepository {
 
     @Override
     public ProductEntity save(ProductEntity product) {
-        return mongoRepository.save(product);
+        try {
+            return mongoRepository.save(product);
+        } catch (Exception e) {
+            throw new ProductPersistenceException("Failed to save product: " + product.getClass().getName(), e);
+        }
     }
 
     @Override
     public void deleteById(String id) {
-        mongoRepository.deleteById(id);
+        if (!mongoRepository.existsById(id)) {
+            throw new ProductNotFoundException(id);
+        }
+
+        try {
+            mongoRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new ProductDeletionException("Failed to delete product with ID: " + id, e);
+        }
     }
 
     @Override
     public Optional<ProductEntity> findById(String id) {
-        return mongoRepository.findById(id);
+        return Optional.ofNullable(
+                mongoRepository.findById(id)
+                        .orElseThrow(() -> new ProductNotFoundException(id))
+        );
     }
 
     @Override
     public List<ProductEntity> findAll() {
-        return mongoRepository.findAll();
+        List<ProductEntity> products = mongoRepository.findAll();
+        if (products.isEmpty()) {
+            throw new ProductsNotFoundException();
+        }
+        return products;
     }
 
     @Override
     public List<ProductEntity> findByCategory(CategoryEnum category) {
-        return mongoRepository.findByCategory(category);
+        List<ProductEntity> products = mongoRepository.findByCategory(category);
+        if (products.isEmpty()) {
+            throw new ProductsByCategoryNotFoundException(category);
+        }
+        return products;
     }
 }
